Moving bindings to avoid circular imports
